When splitting, we reset... Of course the file does not exist so check first

+++ src/Storage/GroupExecutions.php
@@ -70,7 +70,9 @@ class GroupExecutions
      */
     public function delete()
     {
-        unlink($this->pathname);
+        if (file_exists($this->pathname)) {
+            unlink($this->pathname);
+        }
     }
 
     /**
